requirements.txt
asgiref==3.6.0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-countries==7.5.1
django-extensions==3.2.1
django-filter==22.1
django-htmx==1.14.0
django-model-utils==4.3.1
django-phonenumber-field==7.0.2
django-widget-tweaks==1.4.12
folium==0.14.0
gunicorn==20.1.0
humanize==4.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
phonenumbers==8.13.7
phonenumberslite==8.13.7
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.6
python-decouple==3.8
requests==2.28.2
simplejson==3.18.3
sorl-thumbnail==12.9.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0
